# Performance: Introduce Rack::SimpleBodyProxy