--- a/lib/rack/response.rb
+++ b/lib/rack/response.rb
@@ -3,6 +3,7 @@
 require 'rack/request'
 require 'rack/utils'
 require 'rack/body_proxy'
+require 'rack/simple_body_proxy'
 require 'rack/media_type'
 require 'time'
 
@@ -68,7 +69,11 @@ module Rack
         close
         [status.to_i, header, []]
       else
-        [status.to_i, header, BodyProxy.new(self){}]
+        if @block.nil?
+          [status.to_i, header, SimpleBodyProxy.new(@body)]
+        else
+          [status.to_i, header, BodyProxy.new(self){}]
+        end
       end
     end
     alias to_a finish           # For *response