--- a/lib/rack/simple_body_proxy.rb
+++ b/lib/rack/simple_body_proxy.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Rack
+  class SimpleBodyProxy
+    def initialize(body)
+      @body = body
+    end
+
+    def each(&blk)
+      @body.each(&blk)
+    end
+  end
+end